requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2025.6
Django==6.0.2
django-crispy-forms==2.5
django-environ==0.12.0
pillow==12.1.0
psycopg2-binary==2.9.11
sorl-thumbnail==13.0.0
sqlparse==0.5.5
